added service + alert to write smt down
